Handle 404 in HTTPDownload, add a test for license warning, add diagnostics to server test
